--- a/src/MutableDate.php
+++ b/src/MutableDate.php
@@ -105,7 +105,9 @@ class MutableDate extends DateTime implements ChronosInterface
         if ($tz !== $testNow->getTimezone()) {
             $testNow = $testNow->setTimezone($tz ?? date_default_timezone_get());
         }
-        $testNow = $testNow->modify($time);
+        if (!empty($time)) {
+            $testNow = $testNow->modify($time);
+        }
 
         $time = $testNow->format('Y-m-d 00:00:00');
         parent::__construct($time, new DateTimeZone('UTC'));