--- a/tests/Date/ConstructTest.php
+++ b/tests/Date/ConstructTest.php
@@ -13,9 +13,12 @@ declare(strict_types=1);
  */
 namespace Cake\Chronos\Test\Date;
 
+use Cake\Chronos\Chronos;
 use Cake\Chronos\Date;
 use Cake\Chronos\MutableDate;
 use Cake\Chronos\Test\TestCase;
+use DateTimeImmutable;
+use DateTimeZone;
 
 /**
  * Test constructors for Date objects.
@@ -27,18 +30,6 @@ class ConstructTest extends TestCase
      * @return void
      */
     public function testCreateFromTimestamp($class)
-    {
-        $ts = 1454284800;
-        $date = $class::createFromTimestamp($ts);
-        $this->assertEquals('UTC', $date->tzName);
-        $this->assertEquals('2016-02-01', $date->format('Y-m-d'));
-    }
-
-    /**
-     * @dataProvider dateClassProvider
-     * @return void
-     */
-    public function testCreateFromTimestampUtc($class)
     {
         $ts = 1454284800;
         $date = $class::createFromTimestamp($ts);
@@ -258,5 +249,110 @@ class ConstructTest extends TestCase
     {
         $c = new $class('+7 days');
         $this->assertEquals('00:00:00', $c->format('H:i:s'));
+
+        $c = new $class('+10 minutes');
+        $this->assertEquals('00:00:00', $c->format('H:i:s'));
+
+        $c = new $class('2001-01-01 +7 days');
+        $this->assertEquals('2001-01-08', $c->format('Y-m-d'));
+    }
+
+    /**
+     * @dataProvider dateClassProvider
+     */
+    public function testConstructWithLocalTimezone($class)
+    {
+        $londonTimezone = new DateTimeZone('Europe/London');
+
+        // now adjusted to London time
+        // This test could have different results depending on when now is
+        $c = new $class('now', $londonTimezone);
+        $london = new DateTimeImmutable('now', $londonTimezone);
+        $this->assertEquals($london->format('Y-m-d'), $c->format('Y-m-d'));
+        $this->assertSame('UTC', $c->tzName);
+
+        // Confirm that timezone is adjusted before midnight is set.
+        // This implies that the relative time used by Date helpers
+        // is affected by the timezone parameter properly.
+        $london = new DateTimeImmutable('midnight', $londonTimezone);
+        $this->assertEquals('00:00:00', $london->format('H:i:s'));
+
+        $c = $class::today($londonTimezone);
+        $this->assertEquals(Chronos::today($londonTimezone)->format('Y-m-d'), $c->format('Y-m-d'));
+
+        // London timezone is used instead of local timezone
+        $c = new $class('2001-01-02 01:00:00', $londonTimezone);
+        $this->assertEquals('2001-01-02 00:00:00', $c->format('Y-m-d H:i:s'));
+        $this->assertSame('UTC', $c->tzName);
+
+        // London timezone is ignored when timezone is provided in time string
+        $c = new $class('2001-01-01 23:00:00-400', $londonTimezone);
+        $this->assertEquals('2001-01-01 00:00:00', $c->format('Y-m-d H:i:s'));
+        $this->assertSame('UTC', $c->tzName);
+
+        // London timezone is ignored when DateTimeInterface instance is provided
+        $c = new $class(new DateTimeImmutable('2001-01-01 23:00:00-400'), $londonTimezone);
+        $this->assertEquals('2001-01-01 00:00:00', $c->format('Y-m-d H:i:s'));
+        $this->assertSame('UTC', $c->tzName);
+    }
+
+    /**
+     * @dataProvider dateClassProvider
+     */
+    public function testConstructWithLocalTimezoneTestNow($class)
+    {
+        $class::setTestNow(new Chronos('2010-01-01 23:00:00'));
+
+        $londonTimezone = new DateTimeZone('Europe/London');
+
+        // TestNow is adjusted to London time
+        $c = new $class('now', $londonTimezone);
+        $this->assertEquals('2010-01-02 00:00:00', $c->format('Y-m-d H:i:s'));
+        $this->assertSame('UTC', $c->tzName);
+
+        // TestNow is adjusted to London time
+        $c = new $class('+2 days', $londonTimezone);
+        $this->assertEquals('2010-01-04 00:00:00', $c->format('Y-m-d H:i:s'));
+        $this->assertSame('UTC', $c->tzName);
+
+        // TestNow is adjusted to London time
+        $c = $class::today($londonTimezone);
+        $this->assertEquals('2010-01-02 00:00:00', $c->format('Y-m-d H:i:s'));
+        $this->assertEquals(Chronos::today($londonTimezone)->format('Y-m-d'), $c->format('Y-m-d'));
+        $this->assertSame('UTC', $c->tzName);
+
+        // TestNow is ajusted to London time
+        $c = $class::tomorrow($londonTimezone);
+        $this->assertEquals('2010-01-03 00:00:00', $c->format('Y-m-d H:i:s'));
+        $this->assertEquals(Chronos::tomorrow($londonTimezone)->format('Y-m-d'), $c->format('Y-m-d'));
+        $this->assertSame('UTC', $c->tzName);
+
+        // TestNow is ignored when specific date is provided
+        $c = new $class('2001-01-05 01:00:00', $londonTimezone);
+        $this->assertEquals('2001-01-05 00:00:00', $c->format('Y-m-d H:i:s'));
+        $this->assertSame('UTC', $c->tzName);
+    }
+
+    /**
+     * This tests with a large difference between local timezone and
+     * timezone provided as parameter.  This is to help guarantee a date
+     * change would occur so the tests are more consistent.
+     *
+     * @dataProvider dateClassProvider
+     */
+    public function testConstructWithLargeTimzoneChange($class)
+    {
+        $savedTz = date_default_timezone_get();
+        date_default_timezone_set('Pacific/Kiritimati');
+
+        $samoaTimezone = new DateTimeZone('Pacific/Samoa');
+
+        // Pacific/Samoa -11:00 is used intead of local timezone +14:00
+        $c = $class::today($samoaTimezone);
+        $Samoa = new DateTimeImmutable('now', $samoaTimezone);
+        $this->assertEquals($Samoa->format('Y-m-d'), $c->format('Y-m-d'));
+        $this->assertSame('UTC', $c->tzName);
+
+        date_default_timezone_set($savedTz);
     }
 }