https://github.com/cakephp/chronos/issues/149

This allows specifying the timezone for a time passed to Date and MutableDate constructors.  The timezone is used to grab the date before storing it in UTC as before. 

With this change Chronos::today(timezone) and Date::today(timezone) will match (assuming time() hasn't adjusted past a date change of course).

This also fixes a couple missing test case issues where the time was not set to 00:00:00 for Date instances such as using a fixed date + relative time: '2001-01-01 +10 minutes'

The timezone for TestNow instances is also adjusted before the relative time change to support things like 'midnight' being correct in the target timezone (e.g. 00:00:00 regardless of timezone).  This matches PHP's DateTime behavior when passing such values with a timezone parameter.