@othercorey 2019-10-31
@markstory Sorry for all the pushes. I was doing a lot of testing.  

I'm getting a strange error in an unrelated test at the moment.  It seems DateTime::diff() is returning 4 weeks for the last month.

```
1) Cake\Chronos\Test\DateTime\DiffTest::testDiffForHumansRelative with data set #2 (Cake\Chronos\Chronos Object (...), Cake\Chronos\Chronos Object (...), '11 months before')
Failed asserting that two strings are identical.
--- Expected
+++ Actual
@@ @@
-'11 months before'
+'10 months before'
```

Was able to complete a test ahead cakephp 4.x branch.  I was missing support for default/null.

There is an "invalid" test in cakephp where I18n\Date assumes UTC time and expects +23hrs is the same day regardless of local time.  This is relying on an implementation detail of Chronos\Date which is now properly working off local time for relative values.

https://github.com/cakephp/cakephp/blob/f505b7a41cfe7dcb30e09cee7bc62e0645974bfd/tests/TestCase/I18n/DateTest.php#L421