@othercorey 2019-11-01
The weird 1 month - 4 weeks issue resolved itself. I think it's an edge case from testing literally at the end of the month (2019-10-31 23:00:00 - 24:00:00).

I'm not sure how to force the tests to re-run, but they pass on my fork.

https://travis-ci.org/othercorey/chronos/builds/605775613