@othercorey 2019-11-06
@markstory I ran the new benchmark suite against this change.  So far, looks like there are no significant cost differences.

**Before**

```
+-------------------------+---------+------+-----+------------+----------+----------+----------+----------+
| subject                 | set     | revs | its | mem_peak   | best     | mean     | mode     | worst    |
+-------------------------+---------+------+-----+------------+----------+----------+----------+----------+
| benchNow                | date    | 1000 | 5   | 2,483,800b | 5.920μs  | 6.262μs  | 6.025μs  | 7.227μs  |
| benchNowTimezone        | date    | 1000 | 5   | 2,483,816b | 5.734μs  | 6.058μs  | 5.874μs  | 6.802μs  |
| benchRelative           | date    | 1000 | 5   | 2,483,800b | 7.753μs  | 7.995μs  | 8.076μs  | 8.167μs  |
| benchRelativeTimezone   | date    | 1000 | 5   | 2,483,816b | 7.952μs  | 8.096μs  | 8.043μs  | 8.340μs  |
| benchFixed              | date    | 1000 | 5   | 2,483,800b | 5.641μs  | 5.736μs  | 5.769μs  | 5.841μs  |
| benchFixedTimezone      | date    | 1000 | 5   | 2,483,816b | 5.645μs  | 5.800μs  | 5.725μs  | 6.057μs  |
```

**After**

```
+-------------------------+---------+------+-----+------------+----------+----------+----------+----------+
| subject                 | set     | revs | its | mem_peak   | best     | mean     | mode     | worst    |
+-------------------------+---------+------+-----+------------+----------+----------+----------+----------+
| benchNow                | date    | 1000 | 5   | 2,484,776b | 5.751μs  | 6.221μs  | 5.934μs  | 7.391μs  |
| benchNowTimezone        | date    | 1000 | 5   | 2,484,792b | 5.820μs  | 5.948μs  | 6.020μs  | 6.055μs  |
| benchRelative           | date    | 1000 | 5   | 2,484,776b | 7.913μs  | 8.004μs  | 8.030μs  | 8.076μs  |
| benchRelativeTimezone   | date    | 1000 | 5   | 2,484,792b | 7.635μs  | 8.044μs  | 7.929μs  | 8.567μs  |
| benchFixed              | date    | 1000 | 5   | 2,484,776b | 5.323μs  | 5.683μs  | 5.743μs  | 5.947μs  |
| benchFixedTimezone      | date    | 1000 | 5   | 2,484,792b | 5.250μs  | 5.681μs  | 5.591μs  | 6.241μs  |
```